requirements.txt
flask==2.0.1
python-dotenv==0.19.0
pdfplumber==0.7.4
openai==1.3.0
redis==4.3.4
Pillow==9.3.0
pdf2image==1.16.3
pytesseract==0.3.10
python-magic==0.4.27
werkzeug==2.0.1
stripe==2.56.0
gunicorn==23.0.0
flask-cors==3.0.10
httpx==0.27.1
sqlalchemy==2.0.23
PyPDF2==3.0.1
